src/main/python/etl-aws/requirements.txt
boto3==1.23.10
elasticsearch==7.10.1
pandas==1.1.5
python-dateutil==2.8.2
pytz==2022.1
